Combine contract-erc20 and contract-erc20-rpc
