+++ template/examples/contract-erc20/tsconfig.json
