--- template/examples/contract-web3-rpc/.gitignore
@@ -1,2 +1,0 @@
-/build/
-/node_modules/
