--- template/examples/contract-web3-rpc/README.md
@@ -1,11 +1,0 @@
-# ERC20 contract creation (RPC)
-
-This directory contains typescript script describing how to use the web3.js SDK to interact with Frontier.
-
-## Installation and Usage
-
-Use `npm i` to install dependencies. To create an ERC20 contract, execute `node_modules/.bin/ts-node create-erc20.ts` while your template node is running in `--dev` mode.
-
-## Expected output
-
-WIP
